# mentor_ai_engine/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mentor AI Engine",
    "description": "Decision-making engine for the AI mentor: goal understanding, intent parsing, roadmap generation, adaptive planning, nudges, weekly reviews, and feedback loop.",
    "version": "0.8.0-phase8"
  },
  "paths": {
    "/roadmap/generate": {
      "post": {
        "tags": [
          "Phase 2: Roadmap Generator"
        ],
        "summary": "Generate Roadmap",
        "description": "Takes a completed LearnerProfile (output of Phase 1) and returns a\nmilestone/topic roadmap (DAG).\n\nReal mode: two-pass generation \u2014 coarse milestones, then per-milestone\ntopic expansion with prerequisite tracking (see app/roadmap/generator.py).\n\nMock mode (MOCK_LLM=true): fixed mock data, same as the original stub.\n\nEvery path records a metric event (see app/metrics.py, GET /metrics) so\nfallback-to-mock rate is visible without needing external monitoring \u2014\nif real generation starts failing more than expected (e.g. Groq free\ntier flakiness), `roadmap_generation_fallback` will show it.\n\nEvery path also saves the returned roadmap into app/roadmap/store.py \u2014\nPhase 4 (`/roadmap/adapt`) looks roadmaps up by `roadmap_id` from there,\nsince QuizResult only carries the id, not the roadmap itself (backend\nowns long-term persistence; this is just this service's working copy).",
        "operationId": "generate_roadmap_roadmap_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRoadmapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roadmap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roadmap/{roadmap_id}": {
      "get": {
        "tags": [
          "Phase 2: Roadmap Generator"
        ],
        "summary": "Get Roadmap",
        "description": "Debug/ops endpoint \u2014 NOT part of the frozen phase API_CONTRACT, same\nstatus as GET /metrics. Lets you inspect this service's in-memory copy\nof a roadmap (e.g. to see what Phase 4 changed after an /adapt call)\nwithout needing backend's persistence layer in the loop.",
        "operationId": "get_roadmap_roadmap__roadmap_id__get",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roadmap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/roadmap/adapt": {
      "post": {
        "tags": [
          "Phase 4: Adaptive Roadmap Engine"
        ],
        "summary": "Adapt Roadmap",
        "description": "Backend calls this after a learner submits a checkpoint quiz.\n\nReal mode: looks up the roadmap in app/roadmap/store.py (populated by\n/roadmap/generate), reads current Phase 3 mastery scores for the\nlearner, and decides per-topic whether to insert a remediation topic\n(weak topic detection combines quiz score + mastery \u2014 see\napp/adaptive/rules.py), whether to mark the milestone's topics\ncompleted (quiz passed), and whether that unlocks anything downstream.\nRemediation topic *content* comes from one targeted LLM call scoped to\njust the flagged topics \u2014 not a full roadmap regeneration.\n\nFalls back to mock output (not a 500) if: MOCK_LLM=true, the\nroadmap_id isn't in this service's in-memory store (different process,\nrestart, or a roadmap this service never generated), or the\nmilestone_id doesn't exist on that roadmap. Every path records a\nmetric event (see app/metrics.py, GET /metrics).",
        "operationId": "adapt_roadmap_roadmap_adapt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizResult"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdaptRoadmapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/interview/message": {
      "post": {
        "summary": "Send Interview Message",
        "description": "Call this with the learner's reply to the mentor's last question.\n- status == \"in_progress\"  -> show `mentor_message` as the next question\n- status == \"complete\"     -> interview is done, `learner_profile` is the\n                               structured output to hand off to Phase 2\n                               (Roadmap Generator).",
        "operationId": "send_interview_message_interview_message_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/interview/{session_id}": {
      "get": {
        "summary": "Get Interview Status",
        "description": "Poll the current state of an interview session (useful for the frontend to resume).",
        "operationId": "get_interview_status_interview__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdaptRoadmapResponse": {
        "properties": {
          "roadmap_id": {
            "type": "string",
            "title": "Roadmap Id"
          },
          "changed": {
            "type": "boolean",
            "title": "Changed"
          },
          "change_summary": {
            "type": "string",
            "title": "Change Summary"
          },
          "updated_roadmap": {
            "$ref": "#/components/schemas/Roadmap"
          }
        },
        "type": "object",
        "required": [
          "roadmap_id",
          "changed",
          "change_summary",
          "updated_roadmap"
        ],
        "title": "AdaptRoadmapResponse"
      },
...
      "GenerateRoadmapRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "learner_profile": {
            "$ref": "#/components/schemas/LearnerProfile"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "learner_profile"
        ],
        "title": "GenerateRoadmapRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterviewMessageRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "message"
        ],
        "title": "InterviewMessageRequest"
      },
      "InterviewMessageResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "in_progress",
              "complete"
            ],
            "title": "Status"
          },
          "mentor_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mentor Message"
          },
          "learner_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LearnerProfile"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status"
        ],
        "title": "InterviewMessageResponse"
      },
      "LearnerProfile": {
        "properties": {
          "goal": {
            "type": "string",
            "title": "Goal",
            "description": "What the learner wants to achieve, e.g. 'job-ready in backend dev'"
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "description": "Tech/coding sub-domain, e.g. 'web dev', 'DSA', 'ML', 'mobile'"
          },
          "current_level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "title": "Current Level"
          },
          "timeline_weeks": {
            "type": "integer",
            "maximum": 104.0,
            "minimum": 1.0,
            "title": "Timeline Weeks"
          },
          "hours_per_week": {
            "type": "integer",
            "maximum": 80.0,
            "minimum": 1.0,
            "title": "Hours Per Week"
          },
          "known_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Known Skills"
          },
          "weak_areas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Areas"
          },
          "preferred_learning_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Learning Style",
            "description": "e.g. 'project-based', 'video', 'reading docs'"
          },
          "motivation_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivation Type",
            "description": "e.g. 'career-switch', 'exam prep', 'curiosity'"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constraints",
            "description": "e.g. 'works full-time', 'exam in March'"
          }
        },
        "type": "object",
        "required": [
          "goal",
          "domain",
          "current_level",
          "timeline_weeks",
          "hours_per_week"
        ],
        "title": "LearnerProfile"
      },
...
      "QuizResult": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "roadmap_id": {
            "type": "string",
            "title": "Roadmap Id"
          },
          "milestone_id": {
            "type": "string",
            "title": "Milestone Id"
          },
          "quiz_id": {
            "type": "string",
            "title": "Quiz Id"
          },
          "score_percent": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score Percent"
          },
          "topic_breakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Topic Breakdown",
            "description": "topic_id -> score_percent on that topic"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "roadmap_id",
          "milestone_id",
          "quiz_id",
          "score_percent"
        ],
        "title": "QuizResult"
      },
      "Roadmap": {
        "properties": {
          "roadmap_id": {
            "type": "string",
            "title": "Roadmap Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/RoadmapMilestone"
            },
            "type": "array",
            "title": "Milestones"
          },
          "progress_percent": {
            "type": "number",
            "title": "Progress Percent",
            "default": 0.0
          },
          "generated_from_profile_version": {
            "type": "string",
            "title": "Generated From Profile Version",
            "default": "v1"
          }
        },
        "type": "object",
        "required": [
          "roadmap_id",
          "user_id",
          "goal",
          "milestones"
        ],
        "title": "Roadmap"
      },
      "RoadmapMilestone": {
        "properties": {
          "milestone_id": {
            "type": "string",
            "title": "Milestone Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "topics": {
            "items": {
              "$ref": "#/components/schemas/RoadmapTopic"
            },
            "type": "array",
            "title": "Topics"
          },
          "checkpoint_quiz_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Quiz Id"
          }
        },
        "type": "object",
        "required": [
          "milestone_id",
          "title",
          "topics"
        ],
        "title": "RoadmapMilestone"
      },
      "RoadmapTopic": {
        "properties": {
          "topic_id": {
            "type": "string",
            "title": "Topic Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "estimated_hours": {
            "type": "integer",
            "title": "Estimated Hours"
          },
          "status": {
            "type": "string",
            "enum": [
              "locked",
              "available",
              "in_progress",
              "completed"
            ],
            "title": "Status",
            "default": "available"
          },
          "prerequisites": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prerequisites",
            "description": "List of topic_ids"
          }
        },
        "type": "object",
        "required": [
          "topic_id",
          "title",
          "description",
          "estimated_hours"
        ],
        "title": "RoadmapTopic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}